% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_selected}
\alias{plot_selected}
\title{Plot just the selected outline (silhouette)}
\usage{
plot_selected(
  selection,
  fill = "#4C78A8",
  border = "grey20",
  size = 0.4,
  title = NULL,
  show_title = FALSE,
  title_size = 18,
  title_position = c("top", "center"),
  title_alpha = 1
)
}
\description{
Plot just the selected outline (silhouette)
}
