% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_with_surroundings}
\alias{plot_with_surroundings}
\title{Plot selected + surrounding up to a distance (km)}
\usage{
plot_with_surroundings(
  selection,
  distance_km = 15,
  fill_selected = "#4C78A8",
  fill_surroundings = "#A0C4FF",
  border = "grey20",
  size = 0.4,
  title = NULL,
  show_title = FALSE,
  title_size = 18,
  title_position = c("top", "center"),
  title_alpha = 1,
  hint_label = NULL,
  hint_cex = 6,
  hint_alpha = 0.15
)
}
\description{
Plot selected + surrounding up to a distance (km)
}
