% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{LA_game}
\alias{LA_game}
\title{Play an interactive terminal round (LA_game)}
\usage{
LA_game(
  area_type = "lad",
  resolution = "BGC",
  nations = c(E = TRUE, N = FALSE, S = FALSE, W = FALSE),
  random_seed = 99,
  buffers_km = c(15, 30, 60),
  n_options = 4,
  options_radius_km = 60
)
}
\description{
Play an interactive terminal round (LA_game)
}
